+++ microecta-generator/test/Spec.hs
@@ -2,6 +2,7 @@ module Main (main) where
 
 import Test.Hspec (hspec)
 
+import qualified Data.ECTA.DatatypeGenSpec
 import qualified Data.ECTA.GenSpec
 import qualified Data.ECTA.IFCExpressionGenSpec
 import qualified Data.ECTA.RankDecodingSpec
@@ -12,6 +13,7 @@ main :: IO ()
 main =
     hspec $ do
         Data.ECTA.GenSpec.spec
+        Data.ECTA.DatatypeGenSpec.spec
         Data.ECTA.RankDecodingSpec.spec
         Data.ECTA.IFCExpressionGenSpec.spec
         Data.ECTA.RecursiveGenSpec.spec